+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - Change the 'linux-neptune' package to be 'linux-steamos'.
 - Change Pacman to use 'wget' for more reliable downloads on slower internet.
 - Add the ability to build winesapOS using a proxy.
+- Change the package manager from Pamac to bauh.
 
 ## 3.1.1 - The Boot Fix Update